keys/rsa: use lower-case package name
On Linux, I had problems importing the package, since the path uses
lower-case 'rsa', while the package internally uses 'rsa'.

I think this was done to avoid clashes with 'crypto/rsa', as in the
file jose_test.go.

If such a clash occurs, it is easier to change the import statement,
for example
  import (
	"crypto/rsa"                                   // stdlib
	Rsa "github.com/dvsekhvalnov/jose2go/keys/rsa" // helpers
  )